// src/main/resources/static/ts/bo/ICardApi.ts
export interface ICardApi {
    Artist: string;
    Set_Name: string;
    Classifications: string;
    Date_Added: string;
    Set_Num: number;
    Color: string;
    Gamemode: string;
    Franchise: string;
    Image: string;
    Cost: number;
    Inkable: boolean;
    Name: string;
    Type: string;
    Lore: number;
    Rarity: string;
    Flavor_Text: string;
    Unique_ID: string;
    Card_Num: number;
    Body_Text: string;
    Willpower: number;
    Card_Variants: string;
    Date_Modified: string;
    Strength: number;
    Set_ID: string;
}